+++ resources/lang/ca/battles.php
@@ -12,6 +12,7 @@ return [
     '1vs1' => 'Els i les estudiants batallaran de manera individual per a no perdre els seus atributs i mantenir l\'honor!',
     'groupVSgroup' => 'Batalla d\'equips, intentaran guanyar la recompensa',
     'allVSmonsters' => 'La classe lluitarà contra monstres i aconseguiran recompenses al véncer-los',
+    'groupVSmonsters' => 'Un equip s\'enfrontarà contra un monstre, intentaran guanyar la recompensa',
     'not_available' => 'Properament ... Estem treballant-hi :D',
     'loss_fail' => 'Pèrdua de punts per errada',
     'gain' => 'Guany del contrincant si ha encertat la pregunta anterior.',