+++ resources/lang/es/battles.php
@@ -11,6 +11,7 @@ return [
     'start' => 'Empezar',
     '1vs1' => '¡Los y las estudiantes batallarán de manera individual para no perder sus atributos y mantener el honor!',
     'groupVSgroup' => 'Batalla de equipos, intentarán ganar la recompensa',
+    'groupVSmonsters' => 'Un equipo se enfrentará contra un monstruo, intentarán ganar la recompensa',
     'allVSmonsters' => 'La clase luchará contra monstruos y conseguirán recompensas al vencerlos',
     'not_available' => 'Próximamente ... Estamos trabajando en ello :D',
     'loss_fail' => 'Pérdida de puntos por fallo',